requirements.txt
anyio==3.7.0
ariadne==0.19.1
ariadne-django==0.3.0
asgiref==3.7.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.1
django-cors-headers==4.0.0
django-simple-history==3.3.0
djangorestframework==3.14.0
docopt==0.6.2
drf-spectacular==0.26.2
exceptiongroup==1.1.1
graphql-core==3.2.3
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.3
pipreqs==0.4.13
pyrsistent==0.19.3
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
requests==2.31.0
sniffio==1.3.0
sqlparse==0.4.4
starlette==0.28.0
typing_extensions==4.6.2
uritemplate==4.1.1
urllib3==2.0.2
yarg==0.1.9
